Completion 1/n

+++ DotnetBackend.Core/Entities/LGA.cs
@@ -6,7 +6,8 @@ namespace DotnetBackend.Core.Entities
     {
         public string Name { get; set; }
 
-        public int StateId { get; set; }
+        public long StateId { get; set; }
         public virtual State State { get; set; }
+        //public virtual ICollection<Customer> Customers { get; set; }
     }
 }
